requirements.txt
flask
flask_cors
flask-limiter
python-dotenv
langchain
langchain_core
langchain_community
langchain_google_genai
langgraph
werkzeug